Add a Drain::is_enabled method to check if a message should be logged.
This can be used to avoid expensive computation of log message arguments
if the message would be ignored anyway.
This is added as a condition in the `log!` macro,
in order to avoid unessciarrily computing the arguments when the level is disabled.

This is the (somewhat nicer) slog alternative to `log` crate's `log_enabled!` macro.
However, it's still generally preferable to use a `FnValue` as it's
more precise and doens't require additional calls.