Document the minimum supported rust version (MSRV)
Currently, the minimum supported rust version (MSRV) is 1.23.

The full policy for changing the version is available on the Wiki.

The shields.io badge is inspired by serde and indexmap project.

Before this commit, the MSRV was unofficially present in the travis.yml file

To be clear, I have not changed the version and I have not
changed any policies.

I have only documented and summarized what has already existed.