Completely rewrite actions using a matrix
This uses a matrix of different rust versions,
where all combinations are checked & tests.

rust: [1.31, stable, nightly]
features: ["", "serde-values dynamic-keys", "serde-values dynamic-keys nothreads"]

This gives 3*3=9 standard test environments.

This is complicated by the fact that our current Minimal Required Rust
Version fails with `cargo check --all-features`, so we have to specify
it seperately in an `include` block

`cargo check` is required to succeed for all combinations
`cargo test` is required to succeed for default features

`cargo clippy` and `cargo fmt --check` are run independently of the
tests, and only on the latest stable. They are not required for the
build to succeed.

This is based on the indexmap code and a quick read through the
documentation