Use 1 trait for every possible feature combination
Before this change, we had to have 4x as many trait definitions
for every possible combination of std/no_std and threads/nothreads.

Instead, we use an internal utility module of 'maybe' traits,
which are automatically enabled/disabled depending on the features.

So `Sync` becomes `maybe::Sync` and `UwnindSafe` becomes
`maybe::UnwindSafe.

This commit fixes the tests for `feature="nothreads"`
(and probably a ton of other bugs too).