Add BytesKind enum as discussed in PR
This uses a doc(hidden) __Nonexhaustive variant for 1.31 compatibility.
On versions >= 1.40 it conditionally enables #[non_exhaustive]

We discussed Value/Stream distinction in the PR. I also added a 'PlainValue'

I also changed the default impl of both 'Value' and 'Stream' to have a leading `0x`.

Instead of seperating every byte with a space, I changed the default to use an underscore.

So formatting `i32::to_le_bytes(-1)` as a 'Value' would be formatted as `0xFFFF_FFFF`

Also it seperates between every 2 bytes instead of every one.
The old default would be to format this as "FF FF FF FF".

Formatting as 'Stream' avoids the underscores,
but keeps the leading `0x`, so it would be `0xFFFF_FFFF`

Formatting as 'PlainValue' leaves out both the `0x` and the underscores.
This would be useful for things like checksums, which are traditionally
printed without leading `0x` or any underscores

I would appreciate your thoughts on these changes :)