Bump Minimum Supported Rust Version to 1.49
This is the minimum version that supports aarch64-unknown-linux-gnu
and aarch64-apple-darwin as tier one and two especially.

As of this writing, the version is 2 years and 9 months old
so it's more than reasonable to use as a minimum version.

It also allows us to eliminate all our current uses of #[rustversion] for
conditional compilation of newer language features.
However, we may have to re-add that macro in the future.

See release notes: https://blog.rust-lang.org/2020/12/31/Rust-1.49.0.html