CI: Require clippy lints to succeed without warnings
This is only required for MSRV and hardcoded rust versions.
It is possible that an upgrade to stable rust will introduce warnings
that would break existing code.

For this reason, we hardcode a recent version of stable rust
to run clippy on in addition to the MSRV, stable, and nightly.