ci: Require successful test execution for all feature combos.
Before, we only required tests to succeed on the default features.
This was necessary because we had CI disabled for a long time
when Travis CI got rid of their free plans.
It took us a while to switch to Github Actions and by then
some code had broken on special feature combinations.

Now, we require all code to pass tests by default.
We can grant specific exceptions on a case-by-case basis.
In particular, this may be needed for the Minimum Supported Rust Version.