Bump version to v2.8.0-beta.1
There are so many significant changes,
that it makes sense to release a beta version
before doing a full release.

However, this beta release is likely higher quality than v2.7,
simply because it has gone through so many refactorings and fixes.
In particular, it uses Rust 2018 and enables Github Actions.