Support core::error::Error where available
This effectively brings std::error::Error no #[no_std] targets.

It would be nice to have a lint to "forbid" direct use
of std::error and core::error types.
The clippy disallowed-types lint is close,
but doesn't recognize my StdError re-import as a distinct type.
This makes logical sense for a "diallowed types" lint,
but it is not the behavior I want.
A disallowed_paths lint might be better ;)