ci: Don't run clippy as part of test workflow
The clippy::clone_double_ref was triggering on versions 1.56, 1.61, and 1.71,
but was not triggering on stable and nightly.
The lint appears to have been very buggy and so I had to add a suppression for these old versions.

However, since the lint was been renamed to clippy::suspicious_double_ref_op,
we were triggering `renamed-and-removed-lints` on recent versions.

This means we would have to suppress `renamed-and-removed-lints` or use version detection.
It is simpler and cleaner to only run clippy on recent versions.