ci: Cache crates.io index
Older versions 1.56, 1.61 have to pull a git repo to get the crates index.
This is extremely slow, so we want to cache it.
Even for newer versions using the sparse index, it is good to cache.

This cache should not affect correctness,
and is safe to use across multiple rust versions.