ci: Use Swatinem/rust-cache action
This is simpler and more effective than our manual implementation.

Unfortunately, this is completely ineffective on nightly,
since it always keyed by the rust version and will not restore from other versions.
While this is sensible behavior for the target directory,
it means we cannot cache the registry index on nightly.
Luckily, nightly rust uses the sparse index by default,
meaning downloading it is fast anyways.