regression fix: restore compatibility with `erased-serde` version 0.3
The recent release `2.8.0` introduced a breaking change by changing the super-traits and method signatures of the `SerdeValue` trait (enabled via the `nested-value` feature). The change was the update of the public dependency `erased-serde` across semver-breaking versions. The previous version is `0.3`, the new version is `0.4`.

This commit fixes this breakage by reverting the `erased-serde` requirement to `0.3`.

- Closes #358